Move Ruby manifest parsing out of analyzer
