+++ lib/bibliothecary/parsers/rubygems.rb
@@ -4,6 +4,7 @@ module Bibliothecary
   module Parsers
     class Rubygems
       include Bibliothecary::Analyser
+      extend Bibliothecary::MultiParsers::RubyManifest
 
       NAME_VERSION = '(?! )(.*?)(?: \(([^-]*)(?:-(.*))?\))?'.freeze
       NAME_VERSION_4 = /^ {4}#{NAME_VERSION}$/
